Refactored dashboard code to increase readability
