+++ streamlit_dashboard/dashboard.py
@@ -1,17 +1,19 @@
 """Runs the streamlit dashboard"""
 
-import logging
-from datetime import datetime, timedelta
+from datetime import timedelta
 import pandas as pd
 import altair as alt
 import streamlit as st
-import requests
-import psycopg2
-from bs4 import BeautifulSoup
-from database_connection import get_connection, get_cursor
 from streamlit_option_menu import option_menu
-import streamlit_shadcn_ui as ui
 from streamlit_card import card
+from database_connection import (get_connection, get_cursor,
+                                 insert_initial_price, insert_into_product,
+                                 insert_into_subscription, insert_into_website,
+                                 get_latest_price, get_product_info,
+                                 create_account, get_user_id, get_product_subscription,
+                                 stop_tracking_product, execute_database_select_query_fetchall)
+from dashboard_etl import (get_html_from_url, get_website_from_url,
+                           scrape_pricing_process, clean_price)
 
 
 st.set_page_config(
@@ -25,279 +27,11 @@ logout_section = st.container()
 product_section = st.container()
 
 
-def get_html_with_age_gate_bypass(url: str) -> bytes:
-    """Handles Steam URLs with age-gates by simulating form submission."""
-    try:
-        session = requests.Session()
-        session.get(url, timeout=20)
-
-        app_id = url.split('/app/')[1].split('/')[0]
-
-        age_gate_data = {
-            "ageDay": "1",
-            "ageMonth": "January",
-            "ageYear": "1990",
-        }
-        bypass_url = f"https://store.steampowered.com/agecheck/app/{app_id}/"
-        session.post(bypass_url, data=age_gate_data, timeout=20)
-
-        response = session.get(url, timeout=20)
-        response.raise_for_status()
-
-        return response.content
-    except requests.exceptions.RequestException as e:
-        logging.error("Error fetching age-gated URL %s: %s", url, e)
-        return None
-
-
-def get_html_from_url(web_page: str) -> bytes:
-    """ Gets the html content from a given URL"""
-    if "store.steampowered.com" in web_page:
-        return get_html_with_age_gate_bypass(web_page)
-
-    try:
-        html = requests.get(web_page, timeout=20)
-    except requests.exceptions.MissingSchema:
-        return "That URL does not exist."
-    except requests.exceptions.ConnectionError:
-        return "Cannot connect to that URL."
-    if html.status_code > 299 or html.status_code < 200:
-        return f"Error: {html.status_code}."
-    return html.content
-
-
-def get_website_from_url(url: str) -> str:
-    """Gets a main website address from a given URL """
-    website_url = url
-    if ".com" in url:
-        website_url = url.split(".com")[0] + ".com"
-    elif ".co.uk" in url:
-        website_url = url.split(".co.uk")[0] + ".co.uk"
-    return website_url
-
-
-def scrape_pricing_process(html_content: bytes, url: str) -> dict:
-    """ Chooses which scraper to use based off of the URL """
-
-    website_url = get_website_from_url(url)
-
-    if "https://store.steampowered.com" in website_url:
-        return scrape_from_steam_html(html_content, url)
-
-    if "https://www.amazon.com" in url or "https://www.amazon.co." in website_url:
-        return scrape_from_amazon_html(html_content, url)
-
-    logging.error(
-        "Cannot scrape that URL, since it's not an Amazon/Steam webpage.")
-    return
-
-
-def scrape_from_amazon_html(html_content: bytes, url: str) -> dict:
-    """Scrapes product, price and website information from Amazon."""
-    s = BeautifulSoup(html_content, 'html.parser')
-
-    results = s.find("div", id="corePriceDisplay_desktop_feature_div")
-
-    if not results:
-        logging.error("Can't scrape from Amazon URL")
-        return None
-
-    product_title_element = s.find(id="productTitle")
-
-    if not product_title_element:
-        logging.error("Cannot find game title on the page for URL: %s", url)
-        return None
-
-    discount_price = results.find(
-        "div", class_="a-section a-spacing-none aok-align-center aok-relative").find("span", class_="aok-offscreen").text
-    original_price = results.find(
-        "div",
-        class_="a-section a-spacing-small aok-align-center").find("span", class_="a-offscreen").text
-    product_title = product_title_element.text.strip()
-
-    product_information = {
-        "original_price": original_price,
-        "discount_price": discount_price,
-        "game_title": product_title,
-        "website": get_website_from_url(url)}
-    return product_information
-
-
-def scrape_from_steam_html(html_content: bytes, url: str) -> dict:
-    """Scrapes product, price and website information from Steam."""
-    s = BeautifulSoup(html_content, 'html.parser')
-
-    results = s.find(id="game_area_purchase")
-
-    if not results:
-        logging.error("Can't scrape that Steam URL.")
-        return None
-
-    original_price_element = results.find(
-        "div", class_="discount_original_price")
-    discount_price_element = results.find(
-        "div", class_="discount_final_price")
-    game_title_element = s.find(
-        id="appHubAppName", class_="apphub_AppName")
-    regular_price_element = s.find("div", class_="game_purchase_price price", attrs={
-        "data-price-final": True})
-
-    image_element = s.find("img", class_="game_header_image_full")
-    description_element = s.find("div", class_="game_description_snippet")
-
-    if not game_title_element:
-        logging.error("Cannot find product title on the page for URL: %s", url)
-        return None
-    game_title = game_title_element.text.strip()
-
-    if image_element:
-        image_url = image_element['src']
-    else:
-        image_url = None
-
-    if description_element:
-        product_description = description_element.text.strip()
-    else:
-        product_description = "No description found."
-
-    if original_price_element and discount_price_element:
-        original_price = original_price_element.text.strip(
-        ) if original_price_element else "N/A"
-        discount_price = discount_price_element.text.strip(
-        ) if discount_price_element else "N/A"
-    elif regular_price_element:
-        original_price = regular_price_element.text.strip()
-        discount_price = original_price
-    else:
-        original_price = "N/A"
-        discount_price = "N/A"
-
-    product_information = {
-        "original_price": original_price,
-        "discount_price": discount_price,
-        "game_title": game_title,
-        "image_url": image_url,
-        "product_description": product_description,
-        "website": get_website_from_url(url)}
-
-    return product_information
-
-
-def clean_price(price_str: str) -> float:
-    """Cleans current/discount price string to float"""
-    try:
-        cleaned_price = float(price_str.replace(
-            "£", "").replace(",", "").strip())
-        return cleaned_price if cleaned_price >= 0 else None
-    except ValueError:
-        return None
-
-
-def get_website_id(website: str) -> int:
-    """Checks if website exists and returns website id"""
-    try:
-        conn = get_connection()
-        cursor = get_cursor(conn)
-        query = "SELECT website_id FROM website WHERE website_name = %s"
-        cursor.execute(query, (website,))
-        result = cursor.fetchone()
-        return result[0] if result else None
-    except psycopg2.Error as e:
-        print(f"Database error: {e}")
-        return None
-
-
-def get_user_id(email: str) -> int:
-    """Checks if user exists and returns user_id"""
-    try:
-        conn = get_connection()
-        cursor = get_cursor(conn)
-        query = "SELECT user_id FROM users WHERE email_address = %s"
-        cursor.execute(query, (email,))
-        result = cursor.fetchone()
-        return result[0] if result else None
-    except psycopg2.Error as e:
-        print(f"Database error: {e}")
-        return None
-
-
-def get_product_id(url: str) -> int:
-    """Checks if product exists and returns product_id"""
-    try:
-        conn = get_connection()
-        cursor = get_cursor(conn)
-        query = "SELECT product_id FROM product WHERE url = %s"
-        cursor.execute(query, (url,))
-        result = cursor.fetchone()
-        return result[0] if result else None
-    except psycopg2.Error as e:
-        print(f"Database error: {e}")
-        return None
-
-
-def get_subscription_id(user_id: str, product_id: int) -> int:
-    """Checks if subscription exists and returns subscription_id"""
-    try:
-        conn = get_connection()
-        cursor = get_cursor(conn)
-        query = "SELECT subscription_id FROM subscription WHERE user_id = %s and product_id = %s"
-        cursor.execute(query, (user_id, product_id,))
-        result = cursor.fetchone()
-        return result[0] if result else None
-    except psycopg2.Error as e:
-        print(f"Database error: {e}")
-        return None
-
-
-def get_product_subscription(user_id) -> list:
-    """Returns all product ids for a user"""
-    try:
-        conn = get_connection()
-        cursor = get_cursor(conn)
-        query = "SELECT product_id FROM subscription WHERE user_id = %s"
-        cursor.execute(query, (user_id,))
-        result = cursor.fetchall()
-        return result if result else None
-    except psycopg2.Error as e:
-        print(f"Database error: {e}")
-        return None
-
-
-def get_product_info(product_id) -> tuple:
-    """Returns all product info"""
-    conn = get_connection()
-    cursor = get_cursor(conn)
-    try:
-        query = """SELECT product_name, url, original_price, product_description, image_url FROM product WHERE product_id = %s"""
-        cursor.execute(query, (product_id,))
-        return cursor.fetchone()
-    except Exception as e:
-        st.error(f"Error getting the latest price: {e}")
-        return None
-
-
-def get_latest_price(product_id) -> float:
-    """Returns the latest price of a product"""
-    conn = get_connection()
-    cursor = get_cursor(conn)
-    try:
-        query = """SELECT price FROM price_changes WHERE product_id = %s
-        AND timestamp = (SELECT MAX(timestamp) FROM price_changes WHERE product_id = %s)"""
-        cursor.execute(query, (product_id, product_id,))
-        return cursor.fetchone()[0]
-    except Exception as e:
-        st.error(f"Error getting the latest price: {e}")
-        return None
-
-
 def display_charts(product_id) -> alt.Chart:
     """Displays charts for a product"""
     if not product_id:
         st.warning("Please select a valid product.")
         return None
-
-    conn = get_connection()
-    cursor = get_cursor(conn)
 
     query = """
     SELECT price, timestamp 
@@ -305,16 +39,8 @@ def display_charts(product_id) -> alt.Chart:
     WHERE product_id = %s
     ORDER BY timestamp
     """
-
-    cursor.execute(query, (product_id,))
-    result = cursor.fetchall()
-
-    cursor.close()
-    conn.close()
-
-    if not result:
-        st.warning(f"No price data available for Product ID {product_id}.")
-        return None
+    result = execute_database_select_query_fetchall(
+        query, (product_id,), f"No price data available for Product ID {product_id}.")
 
     df = pd.DataFrame(result, columns=['Price', 'Date'])
 
@@ -372,7 +98,7 @@ def login(email: str, password: str) -> bool:
     conn = get_connection()
     cursor = get_cursor(conn)
     try:
-        query = f"""
+        query = """
         SELECT EXISTS (
             SELECT 1
             FROM users
@@ -392,100 +118,6 @@ def login(email: str, password: str) -> bool:
         return False
 
 
-def create_account(first_name, last_name, new_email, new_password) -> bool:
-    """Adds new user data to database"""
-    conn = get_connection()
-    cursor = get_cursor(conn)
-    try:
-        query = """INSERT INTO users (first_name, last_name, email_address, password)
-        VALUES (%s, %s, %s, %s);"""
-        cursor.execute(query, (first_name, last_name, new_email, new_password))
-        cursor.close()
-        conn.commit()
-        conn.close()
-        return True
-    except Exception as e:
-        st.error(f"Error inserting into the database (users): {e}")
-        return False
-
-
-def insert_into_website(website: str) -> int:
-    """Inserts new websites into the website table and returns the corresponding website id"""
-    try:
-        conn = get_connection()
-        cursor = get_cursor(conn)
-        if get_website_id(website):
-            return get_website_id(website)
-        else:
-            cursor.execute(
-                "INSERT INTO website (website_name) VALUES (%s);", (website,))
-            cursor.close()
-            conn.commit()
-            conn.close()
-            return get_website_id(website)
-    except Exception as e:
-        st.error(f"Error inserting into the database (website): {e}")
-        return None
-
-
-def insert_into_product(website_id: int, url: str) -> int:
-    """Inserts new products into the product table and returns the corresponding product id"""
-    product_info = scrape_pricing_process(get_html_from_url(url), url)
-    try:
-        conn = get_connection()
-        cursor = get_cursor(conn)
-        if get_product_id(url):
-            return get_product_id(url)
-        else:
-            cursor.execute(
-                """INSERT INTO product (product_name, url, website_id, original_price, image_url, product_description) VALUES (%s, %s, %s, %s, %s, %s);""",
-                (product_info.get("game_title"), url, website_id,
-                 clean_price(product_info.get("original_price")), product_info.get("image_url"), product_info.get("product_description"),))
-            cursor.close()
-            conn.commit()
-            conn.close()
-            return get_product_id(url)
-    except Exception as e:
-        st.error(f"Error inserting into the database (product): {e}")
-        return None
-
-
-def insert_into_subscription(user_id, product_id, notification_price):
-    """Inserts new subscriptions into the subscription table and returns the corresponding product id"""
-    try:
-        conn = get_connection()
-        cursor = get_cursor(conn)
-        if get_subscription_id(user_id, product_id):
-            return get_subscription_id(user_id, product_id)
-        else:
-            cursor.execute(
-                """INSERT INTO subscription (user_id, product_id, notification_price) VALUES (%s, %s, %s);""",
-                (user_id, product_id, notification_price,))
-            cursor.close()
-            conn.commit()
-            conn.close()
-            return get_subscription_id(user_id, product_id)
-    except Exception as e:
-        st.error(f"Error inserting into the database (subscription): {e}")
-        return None
-
-
-def insert_initial_price(price, product_id) -> None:
-    """Inserts initial price data into the price_changes table"""
-    conn = get_connection()
-    cursor = get_cursor(conn)
-    try:
-        cursor.execute(
-            """INSERT INTO price_changes (price, product_id, timestamp) VALUES (%s, %s, %s);""",
-            (price, product_id, datetime.now().strftime("%Y-%m-%d %H:%M:%S")))
-        cursor.close()
-        conn.commit()
-        conn.close()
-    except Exception as e:
-        st.error(f"Error inserting into the database (price_changes): {e}")
-        return None
-
-
 def track_product(user_id, url, notification_price):
     """Inserts product details into the database"""
     website_id = insert_into_website(get_website_from_url(url))
@@ -495,21 +127,6 @@ def track_product(user_id, url, notification_price):
         "discount_price")
     insert_initial_price(clean_price(price), product_id)
     st.toast("New product successfully tracked!")
-
-
-def stop_tracking_product(user_id, product_id):
-    """Unsubscribes user from a product"""
-    conn = get_connection()
-    cursor = get_cursor(conn)
-    try:
-        cursor.execute(
-            """DELETE FROM subscription WHERE user_id = %s AND product_id = %s""", (user_id, product_id,))
-        cursor.close()
-        conn.commit()
-        conn.close()
-    except Exception as e:
-        st.error(f"Error unsubscribing from product tracking: {e}")
-        return None
 
 
 def show_about_page():
@@ -565,6 +182,52 @@ def view_product(product_id, user_id):
             st.toast(f"""You unsubscribed from tracking {product_name}""")
 
 
+def show_track_page():
+    """Displays the new product tracking page"""
+    user_id = st.session_state.get('user_id')
+    st.session_state["Track new products"] = True
+    st.header("Track a new product")
+    url = st.text_input("Enter a new product URL: ")
+    notification_price = st.text_input(
+        label="Enter the price threshold to receive email notifications about price drops: ")
+    st.button("Track", on_click=track_clicked,
+              args=(user_id, url, notification_price))
+
+
+def show_current_products_page():
+    """Displays the current products"""
+    user_id = st.session_state.get('user_id')
+    product_subscriptions = get_product_subscription(user_id)
+    if not product_subscriptions:
+        st.markdown("You are not currently tracking anything!")
+    else:
+        cols = st.columns(3, gap="small")
+        product_info = []
+        for product_id in product_subscriptions:
+            extracted_info = list(get_product_info(product_id))
+            extracted_info.append(product_id[0])
+            product_info.append(extracted_info)
+        for i, product in enumerate(product_info):
+            product_name = product[0]
+            latest_price = get_latest_price(product[-1])
+            image = product[4]
+            with cols[i % 3]:
+                card(
+
+                    title=f"{product_name}",
+                    text=f"""£{latest_price}""",
+                    image=image,
+                    styles={"card": {
+                        "margin": "0px",
+                        "padding": "0px",
+                        "box-shadow": "0 0"
+                    }, "div": {
+                        "background": "#0000000"
+                    }},
+                    on_click=lambda: set_product(product[-1])
+                )
+
+
 def show_main_page():
     """Displays the main page on the dashboard"""
     with main_section:
@@ -572,7 +235,6 @@ def show_main_page():
             page = option_menu(
                 menu_title="Menu", options=["About", "Current products", "Track new products"])
             st.button("Log Out", key="logout", on_click=logged_out_clicked)
-
         user_id = st.session_state.get('user_id')
         if "current_product" in st.session_state:
             view_product(st.session_state.current_product, user_id)
@@ -581,49 +243,13 @@ def show_main_page():
             st.session_state["About"] = True
             show_about_page()
         elif page == "Track new products":
-            st.session_state["Track new products"] = True
-            st.header("Track a new product")
-            url = st.text_input("Enter a new product URL: ")
-            notification_price = st.text_input(
-                label="Enter the price threshold to receive email notifications about price drops: ")
-            st.button("Track", on_click=track_clicked,
-                      args=(user_id, url, notification_price))
+            show_track_page()
         elif page == "Current products":
-            st.session_state["Current products"] = True
-            user_id = st.session_state.get('user_id')
-            product_subscriptions = get_product_subscription(user_id)
-
-            if not product_subscriptions:
-                st.markdown("You are not currently tracking anything!")
-            else:
-                cols = st.columns(3, gap="small")
-                product_options = [
-                    [get_product_info(product_id), product_id] for product_id in product_subscriptions
-                ]
-                for i, product in enumerate(product_options):
-                    product_name = product[0][0]
-                    original_price = product[0][2]
-                    latest_price = get_latest_price(product[1])
-                    image = product[0][4]
-
-                    with cols[i % 3]:
-                        card(
-
-                            title=f"{product_name}",
-                            text=f"""£{latest_price}""",
-                            image=image,
-                            styles={"card": {
-                                "margin": "0px",
-                                "padding": "0px",
-                                "box-shadow": "0 0"
-                            }, "div": {
-                                "background": "#0000000"
-                            }},
-                            on_click=lambda: set_product(product[1])
-                        )
+            show_current_products_page()
 
 
 def set_product(product_id):
+    """Changes session state to current product"""
     st.session_state["current_product"] = product_id
     st.rerun()
